requirements.txt
Flask==3.0.3
pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
segmentation-models==1.0.1
tensorflow==2.16.1
pytest==8.2.0
azure-monitor-opentelemetry==1.4.2
